Added Suggested Changes.

+++ src/Components/HomePage.tsx
@@ -4,9 +4,22 @@ import { Link, useNavigate } from 'react-router-dom'
 import Navbar from './Navbar'
 import ScrollToTop from 'react-scroll-to-top'
 
+interface productProps {
+  id: number,
+  title: string,
+  description: string,
+  images: string[],
+  thumbnail: string
+}
+
+interface HomeProps {
+  products: productProps[] | undefined;
+  getProducts: () => Promise<null | undefined>;
+  handleDelete: (id: number) => Promise<null | undefined>;
+}
 const HomePage = () => {
   const context = useContext(productContext)
-  const { products, getProducts, handleDelete }: any = context
+  const { products, getProducts, handleDelete }: HomeProps = context
   const navigate = useNavigate()
 
 
@@ -23,12 +36,18 @@ const HomePage = () => {
   const getProductsFn = async () => {
     const res = await getProducts()
     if (res === null) navigate('/login')
-  }
-
+    }
+  
   const handleDeleteItem = async (id: number) => {
-    const res = await handleDelete(id)
-    if (res === null) navigate('/login')
-    else alert('Item deleted successfully')
+    const del = confirm('Are you sure want to delete')
+    if(del === true) {
+      const res = await handleDelete(id)
+      if (res === null) navigate('/login')
+      else alert('Item deleted successfully')
+    }
+    else {
+      alert("Item is not deleted")
+    }
   }
   return (
     <>
@@ -36,7 +55,7 @@ const HomePage = () => {
       <Navbar />
       <div className='container mt-3'>
         <div className='row'>
-          {products.map((product: any, ind: number) => {
+          {products && products.map((product: productProps, ind: number) => {
             return <>
               <div key={product.id} className='col-md-3 my-3 text-center'>
                 <div className="card">
@@ -45,7 +64,7 @@ const HomePage = () => {
                     <h5 className="card-title">{product.title.slice(0, 20) + (product.title.length > 20 ? "..." : "")}</h5>
                     <p className="card-text">{product.description.slice(0, 20) + (product.description.length > 20 ? "..." : "")}</p>
 
-                    <Link to={`product/${product.id}`} className="btn btn-dark">Read More</Link>
+                    <Link to={`product/${ind + 1}`} className="btn btn-dark">Read More</Link>
 
                     <button className="fa-solid fa-trash-can bg-light ms-3 btn btn-light" onClick={() => handleDeleteItem(ind)}></button>
                   </div>
